feat(worker-utils): create getTransferListForWriter function
